Bone-Fracture-Detection-master/requirements_loose.txt
customtkinter
PyAutoGUI
PyGetWindow
Pillow
numpy<2
tensorflow
pandas
matplotlib
scikit-learn
colorama
